Line_bot/本番/requirements.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
future==0.18.2
google-api-core==1.31.2
google-api-python-client==2.15.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
gspread==4.0.1
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyDrive==1.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
